app/requirements.txt
attrs==20.1.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
oauthlib==3.1.0
packaging==20.4
pdfminer.six==20201018
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.0.1
python-dotenv==0.14.0
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sortedcontainers==2.3.0
toml==0.10.1
urllib3==1.26.2
Werkzeug==1.0.1
